+++ apps/improve/src/app/modules/categories/pages/random/random.component.ts
@@ -1,10 +1,5 @@
-import {
-  ChangeDetectionStrategy,
-  ChangeDetectorRef,
-  Component,
-} from '@angular/core';
-import { CategoriesService, ICategory } from '../../../database/public-api';
-import { CurrentClassService } from '../../../database/public-api';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { CategoriesService } from '../../../database/public-api';
 import { ItemsHelperService, ITEMS_SERVICE } from '../../../shared/public-api';
 
 @Component({
@@ -17,27 +12,4 @@ import { ItemsHelperService, ITEMS_SERVICE } from '../../../shared/public-api';
     ItemsHelperService,
   ],
 })
-export class RandomComponent {
-  /**
-   * @internal
-   */
-  public category?: ICategory;
-
-  constructor(
-    private itemsHelper: ItemsHelperService<ICategory>,
-    private currentClass: CurrentClassService,
-    private cdr: ChangeDetectorRef
-  ) {
-    this.category = this.getRandom();
-  }
-
-  public onNewCategory() {
-    this.category = this.getRandom();
-    this.cdr.detectChanges();
-  }
-
-  private getRandom(): ICategory {
-    const result = this.itemsHelper.random(this.currentClass.get());
-    return result.item;
-  }
-}
+export class RandomComponent {}