+++ apps/improve/src/app/modules/themes/pages/random/random.component.ts
@@ -1,19 +1,6 @@
-import {
-  ChangeDetectionStrategy,
-  ChangeDetectorRef,
-  Component,
-} from '@angular/core';
-import {
-  ThemesService,
-  UsedThemesService,
-  CurrentClassService,
-  ITheme,
-} from '../../../database/public-api';
-import {
-  ItemsHelperService,
-  ITEMS_SERVICE,
-  USED_SERVICE,
-} from '../../../shared/public-api';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ThemesService, UsedThemesService } from '../../../database/public-api';
+import { ITEMS_SERVICE, USED_SERVICE } from '../../../shared/public-api';
 
 @Component({
   selector: 'improve-themes-random',
@@ -23,32 +10,6 @@ import {
   providers: [
     { provide: ITEMS_SERVICE, useExisting: ThemesService },
     { provide: USED_SERVICE, useExisting: UsedThemesService },
-    ItemsHelperService,
   ],
 })
-export class RandomComponent {
-  /**
-   * @internal
-   */
-  public theme?: ITheme;
-
-  constructor(
-    private itemsHelper: ItemsHelperService<ITheme>,
-    private currentClass: CurrentClassService,
-    private cdr: ChangeDetectorRef
-  ) {
-    this.theme = this.getRandom();
-  }
-
-  public onNewTheme() {
-    this.theme = this.getRandom();
-    this.cdr.detectChanges();
-  }
-
-  private getRandom(): ITheme {
-    const result = this.itemsHelper.random(this.currentClass.get());
-    this.itemsHelper.setAsUsed(this.currentClass.get(), result.index);
-
-    return result.item;
-  }
-}
+export class RandomComponent {}